+++ docs/data-sources/routing_utilization_label.md
@@ -1,0 +1,30 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "genesyscloud_routing_utilization_label Data Source - terraform-provider-genesyscloud"
+subcategory: ""
+description: |-
+  Data source for Genesys Cloud Routing Utilization Labels. Select a label by name.
+---
+
+# genesyscloud_routing_utilization_label (Data Source)
+
+Data source for Genesys Cloud Routing Utilization Labels. Select a label by name.
+
+## Example Usage
+
+```terraform
+data "genesyscloud_routing_utilization_label" "red_label" {
+  name = "red"
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `name` (String) Label name.
+
+### Read-Only
+
+- `id` (String) The ID of this resource.
